requirements.txt
alabaster==0.7.10
configparser==3.5.0
Django==1.10
django-braces==1.12.0
django-ckeditor==5.4.0
django-cors-headers==1.3.1
django-crontab==0.7.1
django-hosts==2.0
django-js-asset==1.0.0
django-oauth-toolkit==0.12.0
django-pipeline==1.6.13
django-redis==4.8.0
django-sortedm2m==1.5.0
django-sslserver==0.19
django-templatetag-handlebars==1.3.1
djangorestframework==3.5.3
docutils==0.13.1
fcm-django==0.2.7
future==0.16.0
futures==3.1.1
geoip2==2.5.0
idna==2.6
imagesize==0.7.1
jdatetime==1.9.0
Jinja2==2.9.6
MarkupSafe==1.0
maxminddb==1.3.0
oauthlib==2.0.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pycountry==17.5.14
pyfcm==1.2.3
Pygments==2.2.0
python-bidi==0.4.0
pytz==2016.10
redis==2.10.5
reportlab==3.4.0
requests==2.13.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
suds-jurko==0.6
urllib3==1.22